+++ src/api/projects/deployments.rs
@@ -11,6 +11,7 @@
 mod create;
 mod delete;
 mod deployment;
+mod deployments;
 mod edit;
 
 pub use self::create::CreateDeployment;
@@ -24,6 +25,12 @@ pub use self::delete::DeleteDeploymentBuilderError;
 pub use self::deployment::Deployment;
 pub use self::deployment::DeploymentBuilder;
 pub use self::deployment::DeploymentBuilderError;
+
+pub use self::deployments::DeploymentOrderBy;
+pub use self::deployments::DeploymentStatusFilter;
+pub use self::deployments::Deployments;
+pub use self::deployments::DeploymentsBuilder;
+pub use self::deployments::DeploymentsBuilderError;
 
 pub use self::edit::DeploymentStatus;
 pub use self::edit::EditDeployment;